a bit of progress to delay exam marking
